Delete README.md
This project took part of the SemEval2024 competition whose purpose was about Emotion Discovery and Reasoning its Flip in Conversation (https://lcs2.in/SemEval2024-EDiReF/). SemEval tasks explore the nature of meaning in natural languages: how to characterize meaning and how to compute it. They encompass a broad range of semantic topics from the lexical level to the discourse level, including word sense identification, semantic parsing, coreference resolution, and sentiment analysis, among others.

We worked on the third sub task which was an EFR in English conversations : Given a dialogue, we aimed to identify the trigger utterance(s) for an emotion-flip in a multi-party conversation dialogue.

That’s why we divided the work into two parts : We first developed a BERT algorithm to correctly classify which emotion was associated with a specific phrase in the conversation. Then we used the same method to detect which sentence in the utterance was considered as a flip reasoning (e.g., a trigger). 

For more details, feel free to check out the paper related to.